0.4.1 - Updated Loader, seperated into different files
